// src/gateways/utils/http_client_web/models/http_client_config.port.ts
export type HttpClientConfig = {
  data?: any
  encoding?: 'utf8' | 'base64'
  headers?: any
  isJson?: boolean
  url?: string
  port?: number
  timeout?: number
  ssl?: boolean
  method?:
  | 'POST'
  | 'GET'
  | 'PUT'
  | 'DELETE'
  responseType?:
  | 'arraybuffer'
  | 'blob'
  | 'document'
  | 'json'
  | 'text'
  | 'stream'
}